# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GetGddlTags200Item = { [key: string]: unknown };

export type SearchGddlLevelsParams = {
name?: string;
/**
 * @minimum 1
 * @maximum 25
 */
limit?: number;
};

export type SearchGddlLevels200 = { [key: string]: unknown };

export type GetGddlLevel200 = { [key: string]: unknown };

export type GetGddlLevelTags200Item = { [key: string]: unknown };